Fixes & JUnit test + Fixed fetching Books by Category return incorrect paginated count + Fixed fetching relevant Publishers missing group by + Fixed delete Publisher also delete associated Books + More repositories JUnit test

+++ back-end/src/main/java/com/ring/bookstore/dtos/dashboard/IStat.java
@@ -1,7 +1,7 @@
 package com.ring.bookstore.dtos.dashboard;
 
 public interface IStat {
-    Integer getTotal();
-    Integer getCurrentMonth();
-    Integer getLastMonth();
+    Double getTotal();
+    Double getCurrentMonth();
+    Double getLastMonth();
 }
